+++ Changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.4.0
+
+- Add dates sorting to **Monthly Report** channel reminder
+- Delete tokens from yml config in favor of using ENV VARs
+
+## 0.3.0
+
+- Humanize **Monthly Report** channel reminder
+
 ## 0.2.0
 
 - Add **Monthly Report** channel reminder